bugfix: replace unsigned with int
Parameters are int in the C++ code, but read as unsigned in Cuda kernels, which is incorrect. This commit replaces all unsigned with int, since comparing a negative int with an unsigned in kernel code coerces the int to unsigned, which is not what I want.

+++ src/nw-gpu4-ml-diagdiag2pass.cu
@@ -11,8 +11,8 @@ __global__ static void Nw_Gpu4_KernelA(
     const int adjcols,
     const int substsz,
     const int indel,
-    const unsigned tileAx,
-    const unsigned tileAy)
+    const int tileAx,
+    const int tileAy)
 {
     extern __shared__ int shmem[/* substsz*substsz + tileAx + tileAy */];
     // the substitution matrix and relevant parts of the two sequences
@@ -134,8 +134,8 @@ __global__ static void Nw_Gpu4_KernelB(
     const int indel,
     const int trows,
     const int tcols,
-    const unsigned tileBx,
-    const unsigned tileBy,
+    const int tileBx,
+    const int tileBy,
     const int d // the current minor tile diagonal in the score matrix (exclude the header row and column)
 )
 {
@@ -363,8 +363,8 @@ NwStat NwAlign_Gpu4_Ml_DiagDiag2Pass(NwParams &pr, NwInput &nw, NwResult &res)
         gridA.y = (int)ceil(float(adjrows) / tileAy);
         gridA.x = (int)ceil(float(adjcols) / tileAx);
         // block dimensions for kernel A
-        unsigned threadsPerBlockA = min(nw.maxThreadsPerBlock, tileAy * tileAx);
-        dim3 blockA{threadsPerBlockA};
+        int threadsPerBlockA = min(nw.maxThreadsPerBlock, tileAy * tileAx);
+        dim3 blockA{(unsigned)threadsPerBlockA};
 
         // calculate size of shared memory per block in bytes
         int shmemsz = (
